Update README.md
Implement a simple application to render 2 files (book file) in an android and iOS application on a tablet. The app must emulate the functionality where you are reading a book file in your app like in a kindle app.

 

Please use the 2 sample text files provided to be rendered on your app. You must be able to show and select which file to render on a book list then render the book file selected. And no need to design or style the app, just show the basic functionality of showing a list of books and rendering a book file into your react-native app or emulating that "reading a book" functionality.

 

Send your repository link of your code so that we can review it and send a video screen recording of running your app in Android and IOS tablet and to show that it is working on your end.  You can use your built-in device emulator to show your app in android and ios that it is working.